requirements.txt
pandas
requests
beautifulsoup4
psycopg[binary]
python-dotenv